http 初稳 v1

+++ src/core/module/http/mvc/Filter.groovy
@@ -1,10 +1,6 @@
 package core.module.http.mvc
 
-import java.lang.annotation.Documented
-import java.lang.annotation.ElementType
-import java.lang.annotation.Retention
-import java.lang.annotation.RetentionPolicy
-import java.lang.annotation.Target
+import java.lang.annotation.*
 
 @Target([ElementType.METHOD])
 @Retention(RetentionPolicy.RUNTIME)
@@ -14,5 +10,5 @@ import java.lang.annotation.Target
      * 优先级 越大越先执行
      * @return
      */
-    float order() default 0f
+    int order() default 0
 }
